requirements.txt
sentence-transformers
pandas
datasets
ipywidgets
accelerate
google-genai
umap-learn
hdbscan
plotly
